feat: support returning DbResult as a JsonDocument for `addItem` and `bulkAddItem` methods.
Currently it's not possible to read the ID of objects inserted in stores via `AddRecordAsync` and friends, which makes retrieval of the new object ID tedious when the store has been set up to use auto incrementing primary keys.

This patch introduces support for encapsulating the Dexie result on add and bulkAdd on the JS side encoded as a JsonDocument, that way we can introspect the returned document and retireve ID tokens if needed.

I intentionally opted for JsonDocument given that `BlazorDbEvent` class is not generic and it's too difficult to add cases for all the serialization rules.

According to https://developer.mozilla.org/en-US/docs/Web/API/IDBCursor/primaryKey the spec says primary key fields can be of any type.